backend/requirements.txt
fastapi
fastapi-cache2[redis]
redis[hiredis]
aiohttp
uvicorn
python-dotenv
